requirements.test.txt
coveralls
django-form-admin
django-ajax-selects
django-autocomplete-light
